pyproject.toml
[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"


[tool.nitpick]
style = "https://raw.githubusercontent.com/wemake-services/wemake-python-styleguide/master/styles/nitpick-style-wemake.toml"


[tool.poetry]
name = "imagehosting"
description = "Image hosting"
version = "0.1.0"
readme = "README.md"
authors = []


[tool.poetry.dependencies]
python = "3.7.7"

# TODO: remove after https://github.com/python-poetry/poetry/issues/2042
importlib-metadata = "<1.2.0,>=1.1.3"

django = "^2.2.11"
django-split-settings = "^1.0"
django-axes = "^5.6"
django-csp = "^3.7"
django-health-check = "^3.11"
django-http-referrer-policy = "^1.0"
django-feature-policy = "^3.4"

psycopg2 = "^2.8"
gunicorn = "^20.0"
python-decouple = "^3.3"
bcrypt = "^3.2"
structlog = "^20.1.0"
typing_extensions = "^3.7.4"
django-extensions = "^3.0.9"
sorl-thumbnail = "^12.6.3"
Pillow = "^7.2.0"
djangorestframework = "^3.11.1"
markdown = "^3.2.2"


[tool.poetry.dev-dependencies]
django-debug-toolbar = "^3.1"
django-querycount = "^0.7"
django-migration-linter = "^2.4"
django-coverage-plugin = "^1.8"
nplusone = "^1.0"

wemake-python-styleguide = "^0.14"
flake8-pytest-style = "^1.3"
flake8-django = "^1.1"
flake8-logging-format = "^0.6"
nitpick = "^0.23"

pytest = "^6.0"
pytest-django = "^3.10"
pytest-cov = "^2.10"
pytest-randomly = "^3.4"
pytest-deadfixtures = "^2.2"
pytest-testmon = "^1.0"
pytest-timeout = "^1.4"
django-test-migrations = "^1.0"
hypothesis = "^5.36"

mypy = "^0.770"
django-stubs = "^1.4"

ipython = "^7.18"

sphinx = "^2.4"
sphinx-autodoc-typehints = "^1.10"
doc8 = "^0.8"

yamllint = "^1.24"
safety = "^1.9"
dotenv-linter = "^0.1"
polint = "^0.4"
dennis = "^0.9"
dump-env = "^1.2"
